+++ js/src/conversations/Details/index.js
@@ -4,7 +4,7 @@ import {withRouter} from 'react-router-dom'
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import * as fas from '@fortawesome/free-solid-svg-icons'
 import {WithContext, Loading, confirm_modal} from 'reactstrap-toolbox'
-import {Editor, empty_editor, has_content, to_markdown} from '../Editor'
+import {Editor, empty_editor, has_content, to_markdown} from '../../Editor'
 import Message from './Message'
 import RightPanel from './RightPanel'
 import Subject from './Subject'
